Finished gst

+++ real.h
@@ -9,7 +9,7 @@ extern REAL *newREAL(double);
 extern double getREAL(REAL *);
 extern double setREAL(REAL *,double);
 extern void displayREAL(void *, FILE *);
-extern double compareREAL(void *,void *);
+extern int compareREAL(void *,void *);
 extern void displayREAL(void *,FILE *);
 extern void freeREAL(void *);
 #endif
